Bug fix: Parse {/var} patterns correctly
Previously, normalizePath() would only split the path on '/', hence
incorrectly dividing '{/var}' patterns: '{' would be left as the last
character of the previous path component, and 'var}' would consequently
not be recognised as a variable component.

This commit fixes this issue by traversing the array after the split and
looking for items ending in '{'. These are then joined with the
subsequent array element and removed from the previous one.