Minor: Add a function loading the complete YAML conf
Router.loadFullConf() allows one to load the complete YAML swagger
configuration, even if it is split in multiple files. Concretely, it
first loads the YAML file, and then traverses the loaded object looking
for 'interfaces' keys. When found, it tries to load their contents if
the members are strings.

Example:
 var router = new Router.Router();
 var conf = router.loadFullConf(yaml_file_name);
 var tree = router.restBaseSpecToTree(conf.spec);