Support full URLs with protocol etc
So far we only supported paths, but not full URLs with protocol, host &
authentication details. We do however need that capability for backend request
templating, so this patch adds it.

The other change introduced by this patch is the option of passing parameters
into the expand() method, rather than updating the property before calling
`.expand()`. This will be cleaner in URI templating situations.