Support passing in options in toString() method
This patch adds support for passing in an options object into the toString()
method. This replaces the previous string 'format' parameter, which will still
work.

This options object object can take parameters, which are then used for
parameter expansion. This expansion now also respects some more URI template
escaping semantics. For example, it is now possible to pass in a path suffix
in a parameter, and have it become an unescaped part of the overall path with
{+var}.

TODO: add tests.