Flatten visitAync calls and release v0.2.7
Before, we were blidly calling P.each() for a node's children even if
there were no children to visit. Moreover, using P.each seems to create
a Promise hierarchy flattened into a long chain, which apparently node
0.10's version of libuv doesn't like (any more).

This commit fixes that by first checking if there's even a need to
decend into a node's children. If there's not, the chain is broken.
Otherwise, each child's call is placed in a separate .then() block.