Add support for {{var}} substitutions
These substitutions don't apply any escaping, and work consistently between
uri & other properties. In the longer term, the goal is to deprecate the use
of {foo} outside of URIs, so that we can eliminate the inconsistent behavior
of that construct between uri & non-uri context.

Task: https://phabricator.wikimedia.org/T118700