Allow the method to be templated as well; release v0.4.5
Allowing the request method to be templated as well can come in handy in
situations where the request method is not known a priori. This is
needed by the service-template in which we want to create a generalised
approach to making requests to RESTBase.

Note that this change is 100% backwards-compatible in that it provides
'get' as the fall-back method and does not alter the return object if
the `mathod` field has not been specified in the template.