Convert all URI path segments to strings
We repeatedly ran into mysterious 'route not found' issues, which were caused
by non-string elements being passed in a `path` parameter to the `URI`
constructor. This issue frequently happens when processing `revision`
parameters, and leads to very subtle failures.

The work-around so far has been to manually make sure that each path segment
is converted to a string, before constructing the URI. This is error-prone and
ugly.

This patch addresses this by always converting UI path parameters to strings,
unless the `asPattern` flag is supplied. The change is simple, and the
performance impact should be minimal. In a benchmark, converting a
four-element array to strings takes about 0.000058ms, so the performance
impact of this change should be below noise level.