Add meta namespace for annotations
So far we used the magic '' node (URL ending in slash) to attach
metadata like API specs. Since we used a regular path, requests for /
would match this metadata node, despite it not necessarily having any
handlers. As a result, handlers registered for a wildcard (ex:
`{+path}`) were never reached.

This patch avoids this interference by separating metadata out into a
separate 'meta' lookup level. At the node level, this is represented by
the use of the 'meta_' prefix. As a result, regular lookups for URLs
ending in a slash do not match metadata-only nodes any more, which means
that wildcard handlers can kick in as expected. Among other things, this
makes it possible to register wildcard handlers at the API root, such as
`/{+path}`.