Support multi-line yaml syntax for expressions
Our templated expressions can become quite hard to read when they
contain merge functions and more complex object literals on a single
line. While YAML supports multi-line syntax in general, we have not been
able to use this so far. The YAML parser we use emits a trailing newline
for multi-line strings, which throws off the expression parser.

This patch addresses this issue by conservatively stripping a single
trailing newline if the template string looks like a templated
expression. As a result, we will be able to properly format longer
expressions for readability, using YAML's built-in multi-line syntax.