Fix route parsing when concrete and wildcard paths overlap
When building a node tree from a set of paths, the normal
child lookup was used, where concrete path segments can match
wildcards. This resulted in unrelated paths being incorrectly
coalesced (or not, depengin on path order). E.g.
    path 1: /{var}/bar
    path 2: /foo/baz
    Expected: { foo: { baz: {} }, '*': { bar: {} } }
    Actual: { '*': { bar: {}, baz: {} } }

Use exact child lookup when building the tree to avoid this.